+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # master
 
+* Do not use `CoreObject`
+* Derive from broccoli-plugin base class
+
+# 0.2.3
+
 * Make `new` operator optional
 * Use [new `.rebuild` API](https://github.com/broccolijs/broccoli/blob/master/docs/new-rebuild-api.md)
 * Avoid mutating array arguments